added shader support, fix a whole bunch of things. add Ngon setting to VWindow, etc

+++ Calculator/fx98603DEngine/Point.cpp
@@ -1,6 +1,6 @@
 #include "Point.h"
 
-Point::Point(int mX, int mY, int mZ)
+Point::Point(int mX, int mY, float mZ)
 {
 	this->x = mX;
 	this->y = mY;
@@ -11,5 +11,5 @@ Point::Point(Vector vecA)
 {
 	this->x = int(vecA(0));
 	this->y = int(vecA(1));
-	this->z = int(vecA(2));
+	this->z = float(vecA(2));
 }
